Set line length to 124

+++ src/app/api/account/route.ts
@@ -12,10 +12,9 @@ export async function GET(request: NextRequest) {
     const rows: AccountTableRow[] = z
       .array(AccountTableSchema)
       .parse(
-        await db.query(
-          'select acct_id, acct_owner, acct_name from accounts where acct_owner = ? order by acct_name',
-          [uid],
-        ),
+        await db.query('select acct_id, acct_owner, acct_name from accounts where acct_owner = ? order by acct_name', [
+          uid,
+        ]),
       )
     return NextResponse.json(rows)
   } catch (e) {
@@ -53,15 +52,9 @@ export async function DELETE(request: NextRequest) {
     const body = await request.json()
     const obj: AccountTableRow = AccountTableSchema.parse(body)
     if (obj.acct_id) {
-      await db.query(
-        `delete from accounts where acct_owner = ? and acct_id = ?`,
-        [uid, obj.acct_id],
-      )
+      await db.query(`delete from accounts where acct_owner = ? and acct_id = ?`, [uid, obj.acct_id])
     } else if (obj.acct_name) {
-      await db.query(
-        `delete from accounts where acct_owner = ? and acct_name = ?`,
-        [uid, obj.acct_name],
-      )
+      await db.query(`delete from accounts where acct_owner = ? and acct_name = ?`, [uid, obj.acct_name])
     }
     return await GET(request)
   } catch (e) {
